The new file '' The new file 'Lesson1-2' is made for all tasks from initial lessons and file 'Program' is set as Main
